+++ lib/core/database/module_database.dart
@@ -1,0 +1,3 @@
+abstract class LunaModuleDatabase {
+    void registerAdapters();
+}