+++ lib/modules/settings/modules/logs.dart
@@ -1,2 +1,0 @@
-export 'logs/route.dart';
-export 'logs/widgets.dart';