+++ lib/modules/settings/modules/logs_details.dart
@@ -1,2 +1,0 @@
-export 'logs_details/route.dart';
-export 'logs_details/widgets.dart';