+++ lib/modules/settings/modules/system.dart
@@ -1,2 +1,1 @@
 export 'system/route.dart';
-export 'system/widgets.dart';