+++ lib/modules/settings/modules/system/widgets.dart
@@ -1,3 +1,0 @@
-export 'widgets/clear_configuration_tile.dart';
-export 'widgets/sentry_tile.dart';
-export 'widgets/version_tile.dart';