+++ lib/modules/settings/modules/system_logs_details.dart
@@ -1,0 +1,2 @@
+export 'system_logs_details/route.dart';
+export 'system_logs_details/widgets.dart';